수업은 완

+++ src/main/java/com/greengram/application/feed/FeedRepository.java
@@ -3,5 +3,5 @@ package com.greengram.application.feed;
 import com.greengram.entity.Feed;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FeedRepository extends JpaRepository<Feed, Integer> {
+public interface FeedRepository extends JpaRepository<Feed, Long> {
 }
